requirements.txt
pytest
requests
allure-pytest
pydantic
pytest-xdist
